fix 0.2

+++ src/Traits/FileSystemModelTrait.php
@@ -64,26 +64,29 @@ trait FileSystemModelTrait
      * @param array data
      * @param array whiteList
      *
-     * @return boolean
+     * @return bool
      */
-    public function update($data = null, $whiteList = null) : bool
+    public function updateOrFail($data = null, $whiteList = null) : bool
     {
         //associate uploaded files
         if (isset($data['files'])) {
             if (!empty($data['files'])) {
                 /**
                  * @todo for now lets delete them all and updated them
-                 * look for a better solution later, this can since we are not using transactions
+                 * look for a better solution later , this can cause issues
+                 * since we are not using transactions
                  */
                 $this->deleteFiles();
 
                 $this->uploadedFiles = $data['files'];
             } else {
-                $this->deleteFiles();
+                if ((bool) $this->di->get('app')->get('delete_images_on_empty_files_field')) {
+                    $this->deleteFiles();
+                }
             }
         }
 
-        return parent::update($data, $whiteList);
+        return parent::updateOrFail($data, $whiteList);
     }
 
     /**
@@ -110,9 +113,9 @@ trait FileSystemModelTrait
      * @param array data
      * @param array whiteList
      *
-     * @return boolean
+     * @return bool
      */
-    public function save($data = null, $whiteList = null) : bool
+    public function saveOrFail($data = null, $whiteList = null) : bool
     {
         //associate uploaded files
         if (isset($data['files'])) {
@@ -121,7 +124,7 @@ trait FileSystemModelTrait
             }
         }
 
-        return parent::save($data, $whiteList);
+        return parent::saveOrFail($data, $whiteList);
     }
 
     /**
@@ -131,12 +134,14 @@ trait FileSystemModelTrait
      */
     public function deleteFiles() : bool
     {
-        $systemModule = SystemModules::getSystemModuleByModelName(self::class);
+        $systemModule = SystemModules::getByModelName(self::class);
 
         if ($files = FileSystemEntities::getAllByEntityId($this->getId(), $systemModule)) {
-            $files->update([], function ($file) {
-                $file->softDelete();
-            });
+            $files->filter(
+                function ($file) {
+                    $file->softDelete();
+                }
+            );
         }
 
         return true;
@@ -166,7 +171,7 @@ trait FileSystemModelTrait
     }
 
     /**
-     * Given the array of files we will attch this files to the files.
+     * Given the array of files we will attach this files to the files.
      * [
      *  'file' => $file,
      *  'file_name' => 'dfadfa'
@@ -361,7 +366,7 @@ trait FileSystemModelTrait
                 ->useCustomMapper($fileMapper);
 
             /**
-             * @todo create a mapper for entity so we dont have to look for the relationship?
+             * @todo create a mapper for entity so we don't have to look for the relationship?
              */
             return $this->di->getMapper()->map($fileEntity, Files::class);
         }
@@ -391,7 +396,7 @@ trait FileSystemModelTrait
     protected function searchCriteriaForFilesByName(string $fieldName) : array
     {
         $systemModule = SystemModules::getSystemModuleByModelName(self::class);
-        $appPublicImages = (bool) $this->di->getApp()->get('public_images');
+        $appPublicImages = (bool) $this->di->get('app')->get('public_images');
 
         $bindParams = [
             'system_module_id' => $systemModule->getId(),
@@ -427,7 +432,7 @@ trait FileSystemModelTrait
     public function getAttachmentByNameAndAttributes(string $fieldName, string $key, string $value)
     {
         $systemModule = SystemModules::getSystemModuleByModelName(self::class);
-        $appPublicImages = (bool) $this->di->getApp()->get('public_images');
+        $appPublicImages = (bool) $this->di->get('app')->get('public_images');
 
         $bindParams = [
             'system_module_id' => $systemModule->getId(),
@@ -473,7 +478,7 @@ trait FileSystemModelTrait
     public function getAttachments(string $fileType = null) : ResultsetInterface
     {
         $systemModule = SystemModules::getSystemModuleByModelName(self::class);
-        $appPublicImages = (bool) $this->di->getApp()->get('public_images');
+        $appPublicImages = (bool) $this->di->get('app')->get('public_images');
 
         $bindParams = [
             'system_module_id' => $systemModule->getId(),
